+++ common/src/main/java/com/alien/common/registry/init/AlienEntityTypes.java
@@ -6,6 +6,7 @@ import com.alien.common.gameplay.entity.living.alien.chestburster.Chestburster;
 import com.alien.common.gameplay.entity.living.alien.ovipositor.Ovipositor;
 import com.alien.common.gameplay.entity.living.alien.ovomorph.Ovomorph;
 import com.alien.common.gameplay.entity.living.alien.parasite.facehugger.Facehugger;
+import com.alien.common.gameplay.entity.living.alien.predalien_chestburster.PredalienChestburster;
 import com.alien.common.gameplay.entity.living.alien.xenomorph.boiler.Boiler;
 import com.alien.common.gameplay.entity.living.alien.xenomorph.crusher.Crusher;
 import com.alien.common.gameplay.entity.living.alien.xenomorph.drone.Drone;
@@ -77,6 +78,12 @@ public class AlienEntityTypes {
         "aberrant_predalien",
         EntityType.Builder.of(Predalien::new, AVPEntityTypes.ALIEN_CATEGORY)
             .sized(0.98f, 3.98f)
+    );
+
+    public static final AVPDeferredHolder<EntityType<PredalienChestburster>> ABERRANT_PREDALIEN_CHESTBURSTER = AVPEntityTypes.register(
+        "aberrant_predalien_chestburster",
+        EntityType.Builder.of(PredalienChestburster::new, AVPEntityTypes.ALIEN_CATEGORY)
+            .sized(0.35f, 0.35f)
     );
 
     public static final AVPDeferredHolder<EntityType<Prowler>> ABERRANT_PROWLER = AVPEntityTypes.register(
@@ -253,6 +260,12 @@ public class AlienEntityTypes {
             .sized(0.98f, 3.98f)
     );
 
+    public static final AVPDeferredHolder<EntityType<PredalienChestburster>> NETHER_PREDALIEN_CHESTBURSTER = AVPEntityTypes.register(
+        "nether_predalien_chestburster",
+        EntityType.Builder.of(PredalienChestburster::new, AVPEntityTypes.ALIEN_CATEGORY)
+            .sized(0.35f, 0.35f)
+    );
+
     public static final AVPDeferredHolder<EntityType<Prowler>> NETHER_PROWLER = AVPEntityTypes.register(
         "nether_prowler",
         EntityType.Builder.of(Prowler::new, AVPEntityTypes.ALIEN_CATEGORY)
@@ -305,6 +318,12 @@ public class AlienEntityTypes {
         "predalien",
         EntityType.Builder.of(Predalien::new, AVPEntityTypes.ALIEN_CATEGORY)
             .sized(0.98f, 3.98f)
+    );
+
+    public static final AVPDeferredHolder<EntityType<PredalienChestburster>> PREDALIEN_CHESTBURSTER = AVPEntityTypes.register(
+        "predalien_chestburster",
+        EntityType.Builder.of(PredalienChestburster::new, AVPEntityTypes.ALIEN_CATEGORY)
+            .sized(0.35f, 0.35f)
     );
 
     public static final AVPDeferredHolder<EntityType<Prowler>> PROWLER = AVPEntityTypes.register(
@@ -419,6 +438,10 @@ public class AlienEntityTypes {
         Services.REGISTRY.registerEntityAttributes(ABERRANT_OVOMORPH, Ovomorph::createOvomorphAttributes);
         Services.REGISTRY.registerEntityAttributes(ABERRANT_PRAETORIAN, Praetorian::createPraetorianAttributes);
         Services.REGISTRY.registerEntityAttributes(ABERRANT_PREDALIEN, Predalien::createPredalienAttributes);
+        Services.REGISTRY.registerEntityAttributes(
+            ABERRANT_PREDALIEN_CHESTBURSTER,
+            PredalienChestburster::createPredalienChestbursterAttributes
+        );
         Services.REGISTRY.registerEntityAttributes(ABERRANT_PROWLER, Prowler::createProwlerAttributes);
         Services.REGISTRY.registerEntityAttributes(ABERRANT_QUEEN, Queen::createQueenAttributes);
         Services.REGISTRY.registerEntityAttributes(ABERRANT_RUNNER, Runner::createRunnerAttributes);
@@ -447,6 +470,10 @@ public class AlienEntityTypes {
         Services.REGISTRY.registerEntityAttributes(NETHER_OVOMORPH, Ovomorph::createOvomorphAttributes);
         Services.REGISTRY.registerEntityAttributes(NETHER_PRAETORIAN, Praetorian::createPraetorianAttributes);
         Services.REGISTRY.registerEntityAttributes(NETHER_PREDALIEN, Predalien::createPredalienAttributes);
+        Services.REGISTRY.registerEntityAttributes(
+            NETHER_PREDALIEN_CHESTBURSTER,
+            PredalienChestburster::createPredalienChestbursterAttributes
+        );
         Services.REGISTRY.registerEntityAttributes(NETHER_PROWLER, Prowler::createProwlerAttributes);
         Services.REGISTRY.registerEntityAttributes(NETHER_QUEEN, Queen::createQueenAttributes);
         Services.REGISTRY.registerEntityAttributes(NETHER_RUNNER, Runner::createRunnerAttributes);
@@ -456,6 +483,7 @@ public class AlienEntityTypes {
         Services.REGISTRY.registerEntityAttributes(OVOMORPH, Ovomorph::createOvomorphAttributes);
         Services.REGISTRY.registerEntityAttributes(PRAETORIAN, Praetorian::createPraetorianAttributes);
         Services.REGISTRY.registerEntityAttributes(PREDALIEN, Predalien::createPredalienAttributes);
+        Services.REGISTRY.registerEntityAttributes(PREDALIEN_CHESTBURSTER, PredalienChestburster::createPredalienChestbursterAttributes);
         Services.REGISTRY.registerEntityAttributes(PROWLER, Prowler::createProwlerAttributes);
         Services.REGISTRY.registerEntityAttributes(QUEEN, Queen::createQueenAttributes);
         Services.REGISTRY.registerEntityAttributes(ROYAL_ABERRANT_ADOLESCENT, Adolescent::createAdolescentAttributes);