+++ common/src/main/java/com/avp/common/constant/AttackDamageConstants.java
@@ -22,6 +22,8 @@ public record AttackDamageConstants() {
 
     public static final float PREDALIEN_ATTACK_DAMAGE = PLAYER_HEALTH * 0.75F;
 
+    public static final float PREDALIEN_CHESTBURSTER_ATTACK_DAMAGE = PLAYER_HEALTH * 0.1F;
+
     public static final float PROWLER_ATTACK_DAMAGE = PLAYER_HEALTH * 0.5F;
 
     public static final float QUEEN_ATTACK_DAMAGE = PLAYER_HEALTH * 2.5F;