+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## ✨ What's New
 - Added predalien.
+- Added predalien chestburster.
 - Marines can now be given armor items.
 - Marines can now use armor items:
   - Marines will now automatically figure out what armor is best to wear depending on their situation.